feat: solve modules
